+++ actions/discounts/delete-discount.ts
@@ -1,9 +1,12 @@
 "use server";
 
+import { rolePermissions, UserRole } from "@/lib/permissions";
 import { prisma } from "@/lib/prisma";
+import { checkUserPermissions } from "../users/check-permissions";
 
 export const deleteDiscount = async (id: string) => {
   try {
+    await checkUserPermissions([...rolePermissions[UserRole.ACCOUNTANT]]);
     await prisma.discount.delete({
       where: { id },
     });
