+++ actions/users/delete-user.ts
@@ -1,6 +1,8 @@
 "use server";
 
+import { rolePermissions, UserRole } from "@/lib/permissions";
 import { prisma } from "@/lib/prisma";
+import { checkUserPermissions } from "./check-permissions";
 
 export const deleteUserById = async (id: string) => {
   try {
@@ -8,7 +10,7 @@ export const deleteUserById = async (id: string) => {
     // if (id === req.user!.id) {
     //   return res.status(400).json({ error: "Cannot delete your own account" });
     // }
-
+    await checkUserPermissions([...rolePermissions[UserRole.MANAGER]]);
     await prisma.user.delete({
       where: { id },
     });
